Preferring to version the Graphql controller.
